pull majority of code out
