+++ .gitignore
@@ -65,3 +65,6 @@ fabric.properties
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
 
+# aje
+**/*.pyc
+
